Create #292 Nim Game
We can look at this problem from the other side that for a given N, we figure that if I will lose the game. Since I will take the first turn, if I loss, it means that my friend and I either takes the same times to move the stones. We can turn this problem to a mathematic problem that (x1+x2+---+xa)+(y1+y2+----+ya) = n. In this equation, "a" means the times my friend and I take to move the stones while x, y both among integer 1, 2 and 3. Thus for this question, we can explain this question like this. The condition is that a is an integer, x&y are among [1,3], for a given a, n is fixed no matter the number of x and y. The question is to find out the value of n. 
Solution: Since for a given a, n is fixed no matter the value of x & y, if a=1, then [1,3]+[1,3] = n, n= 4; if a=2, then [2,6] + [2,6] = n, n =8; So n should be divided by 4. In this sense, if(n%4!=0) I have a chance to win this game.